Change the protocol to OsmAnd Live Tracking web address format
