+++ traccar-client/build.gradle
@@ -34,4 +34,5 @@ android {
 }
 
 dependencies {
+    compile 'com.loopj.android:android-async-http:1.4.7'
 }
