+++ traccar-client/src/main/java/org/traccar/client/activity/AboutActivity.java
@@ -13,12 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.traccar.client;
+package org.traccar.client.activity;
 
 import android.app.Activity;
 import android.content.pm.PackageManager.NameNotFoundException;
 import android.os.Bundle;
 import android.widget.TextView;
+
+import org.traccar.client.R;
 
 public class AboutActivity extends Activity {
