+++ traccar-client/src/main/java/org/traccar/client/provider/PositionProvider.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.traccar.client;
-
-import java.util.Date;
+package org.traccar.client.provider;
 
 import android.content.Context;
 import android.location.Location;
@@ -24,26 +22,57 @@ import android.location.LocationManager;
 import android.location.LocationProvider;
 import android.os.Bundle;
 import android.os.Handler;
-import android.widget.Toast;
+
+import org.traccar.client.R;
+import org.traccar.client.activity.StatusActivity;
 
 public class PositionProvider {
 
     public static final String PROVIDER_MIXED = "mixed";
     public static final long PERIOD_DELTA = 10 * 1000;
     public static final long RETRY_PERIOD = 60 * 1000;
-
-    public interface PositionListener {
-        public void onPositionUpdate(Location location);
-    }
-    
     private final Handler handler;
     private final LocationManager locationManager;
     private final long period;
     private final PositionListener listener;
     private final Context context;
-
+    private final InternalLocationListener fineLocationListener = new InternalLocationListener();
+    private final InternalLocationListener coarseLocationListener = new InternalLocationListener();
     private boolean useFine;
     private boolean useCoarse;
+    private final Runnable updateTask = new Runnable() {
+
+        private long lastTime;
+
+        private boolean tryProvider(String provider) {
+            Location location = locationManager.getLastKnownLocation(provider);
+
+            /*if (location != null) {
+                Toast.makeText(context, "phone: " + new Date() + "\ngps: " + new Date(location.getTime()), Toast.LENGTH_LONG).show();
+            } else {
+                Toast.makeText(context, "no location", Toast.LENGTH_LONG).show();
+            }*/
+
+            if (location != null && location.getTime() != lastTime) {
+                lastTime = location.getTime();
+                listener.onPositionUpdate(location);
+                return true;
+            } else {
+                return false;
+            }
+        }
+
+        @Override
+        public void run() {
+            if (useFine && tryProvider(LocationManager.GPS_PROVIDER)) {
+            } else if (useCoarse && tryProvider(LocationManager.NETWORK_PROVIDER)) {
+            } else {
+                listener.onPositionUpdate(null);
+            }
+            handler.postDelayed(this, period);
+        }
+
+    };
 
     public PositionProvider(Context context, String type, long period, PositionListener listener) {
         handler = new Handler(context.getMainLooper());
@@ -87,42 +116,9 @@ public class PositionProvider {
         locationManager.removeUpdates(coarseLocationListener);
     }
 
-    private final Runnable updateTask = new Runnable() {
-
-        private long lastTime;
-
-        private boolean tryProvider(String provider) {
-            Location location = locationManager.getLastKnownLocation(provider);
-
-            /*if (location != null) {
-                Toast.makeText(context, "phone: " + new Date() + "\ngps: " + new Date(location.getTime()), Toast.LENGTH_LONG).show();
-            } else {
-                Toast.makeText(context, "no location", Toast.LENGTH_LONG).show();
-            }*/
-            
-            if (location != null && location.getTime() != lastTime) {
-                lastTime = location.getTime();
-                listener.onPositionUpdate(location);
-                return true;
-            } else {
-                return false;
-            }
-        }
-
-        @Override
-        public void run() {
-            if (useFine && tryProvider(LocationManager.GPS_PROVIDER)) {
-            } else if (useCoarse && tryProvider(LocationManager.NETWORK_PROVIDER)) {
-            } else {
-                listener.onPositionUpdate(null);
-            }
-            handler.postDelayed(this, period);
-        }
-
-    };
-
-    private final InternalLocationListener fineLocationListener = new InternalLocationListener();
-    private final InternalLocationListener coarseLocationListener = new InternalLocationListener();
+    public interface PositionListener {
+        public void onPositionUpdate(Location location);
+    }
 
     private class InternalLocationListener implements LocationListener {
