+++ traccar-client/traccar-client.iml
@@ -86,5 +86,6 @@
     </content>
     <orderEntry type="jdk" jdkName="Android API 22 Platform" jdkType="Android SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" exported="" name="android-async-http-1.4.7" level="project" />
   </component>
 </module>
